requirements.txt
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
google-api-core==1.1.1
google-api-python-client==1.6.6
google-auth==1.4.1
google-cloud-texttospeech==0.1.0
googleapis-common-protos==1.5.3
grpcio==1.11.0
httplib2==0.11.3
idna==2.6
oauth2client==4.1.2
Pillow==5.1.0
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
SpeechRecognition==3.8.1
uritemplate==3.0.0
urllib3==1.22
